Moved other components to components dir
