+++ client/__tests__/subapp.spec.tsx
@@ -1,14 +1,16 @@
-import React from 'react';
-import { SubApp } from '../app/subapp';
-import { shallow } from 'enzyme';
+import React from "react";
+import { SubApp } from "../app/components/subapp";
+import { shallow } from "enzyme";
 
 test("renders the heading with props supplied", () => {
-    const result = shallow(<SubApp message="Polaris"/>).contains(<h1>Hello Polaris</h1>);
-    expect(result).toBeTruthy();
+  const result = shallow(<SubApp message="Polaris" />).contains(
+    <h1>Hello Polaris</h1>
+  );
+  expect(result).toBeTruthy();
 });
 
 test("matches snapshot test with no props supplied", () => {
-    const snap = shallow(<SubApp message=""/>);
-    // Snapshot demo
-    expect(snap).toMatchSnapshot();
- });
+  const snap = shallow(<SubApp message="" />);
+  // Snapshot demo
+  expect(snap).toMatchSnapshot();
+});
